Exiling all graveyards abstracted out into common effect. Added Morningtide. Implementation of Rest In Peace and Relic of Progenitus abstracted out.
